fix: github ci/cd config

+++ scripts/clear-cache.js
@@ -6,5 +6,7 @@ const fs = require("fs")
 const path_node_modules = path.join(__dirname, '..', 'node_modules')
 const path_cache_node_modules = path.join(path_node_modules, ".cache")
 
-fs.rmSync(path_cache_node_modules, { recursive: true })
+if (fs.existsSync(path_cache_node_modules)) {
+    fs.rmSync(path_cache_node_modules, { recursive: true })
+}
